bug fix in airflow kernel function
In von der Lippe et. al 2013, parameters a and b for the lognormal function seem to be switched. The values for "a" seem to be actually values of "b", and viceversa. They provide negative values for "b", which would cause the lognormal to return values of probability <0. Using their values of "a" as "b"and viceversa i their lognormal function allows to replicate (almost) the kernels reported in fig. 3.
This was included in the model by leaving the kernels and configFile as they were, and switching the parameters when calling f_airflow() in SpreadModel(). Comments have been added in both functions.